third_party/blink/web_tests/fast/dom/shadow/shadow-root-blur-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This tests the blur() on on shadow host should work when a shadow host contains a focused element in its shadow DOM subtrees property (bug 81102)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


getNodeInComposedTree('shadowHostA/childInTreeOne').focus()
PASS document.activeElement.id is "shadowHostA"
getNodeInComposedTree('shadowHostA').blur();
shadowHostA onblur called!
PASS document.activeElement is document.body
PASS getNodeInComposedTree('shadowHostA/').activeElement is null
PASS getNodeInComposedTree('shadowHostA/shadowHostB/').activeElement is null
PASS getNodeInComposedTree('shadowHostC/').activeElement is null
getNodeInComposedTree('shadowHostA/shadowHostB/childInTreeTwo').focus()
PASS document.activeElement.id is "shadowHostA"
getNodeInComposedTree('shadowHostA/shadowHostB').blur();
shadowHostB onblur called!
shadowHostA onblur called!
PASS document.activeElement is document.body
PASS getNodeInComposedTree('shadowHostA/').activeElement is null
PASS getNodeInComposedTree('shadowHostA/shadowHostB/').activeElement is null
PASS getNodeInComposedTree('shadowHostC/').activeElement is null
getNodeInComposedTree('shadowHostA/shadowHostB/childInTreeTwo').focus()
PASS document.activeElement.id is "shadowHostA"
getNodeInComposedTree('shadowHostA').blur();
shadowHostB onblur called!
shadowHostA onblur called!
PASS document.activeElement is document.body
PASS getNodeInComposedTree('shadowHostA/').activeElement is null
PASS getNodeInComposedTree('shadowHostA/shadowHostB/').activeElement is null
PASS getNodeInComposedTree('shadowHostC/').activeElement is null
getNodeInComposedTree('shadowHostC/childInTreeThree').focus()
PASS document.activeElement.id is "shadowHostC"
getNodeInComposedTree('shadowHostA').blur();
PASS document.activeElement.id is "shadowHostC"
PASS getNodeInComposedTree('shadowHostA/').activeElement is null
PASS getNodeInComposedTree('shadowHostA/shadowHostB/').activeElement is null
PASS getNodeInComposedTree('shadowHostC/').activeElement.id is "childInTreeThree"
PASS successfullyParsed is true

TEST COMPLETE
